Use raw axis input for determining dash direction
Avoids the acceleration lag when performing a dash. If the normal ('GetAxis') smoothed input is used, it's possible to press a direction key and then dash quickly enough that the dash goes in the wrong direction. This change means that you can't dash in arbitrary fractions of a diagonal (because we aren't smoothing between directions anymore), but it wasn't a reliable way to dash anyway.